requirements.txt
cloudevents
dapr
dapr.ext.grpc
hera-workflows
minio
pyyaml
requests